refactor(audio): extract content filtering into content filter module
remove duplicated filterContentForAudio from HybridAudioPlayer.tsx and
private extractReadableContent/extractFromHTML/extractFromMarkdown/cleanText
methods from audioPlayer.ts; consolidate into src/lib/contentFilter.ts
